server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
waitress~=2.1.2
requests~=2.31.0
PyJWT==2.1.0
uuid
pymongo
Werkzeug==2.2.2
pandas